Revise python tutorial lesson 16 to make sure tutorial text and code match.
